Подготовка к pr

+++ main-service/src/main/java/ru/practicum/dto/event/EventShortDto.java
@@ -23,5 +23,5 @@ public class EventShortDto {
     Boolean paid;
     String title;
     Integer views;
-    Integer comments;
+    Long comments;
 }
